requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
cssselect==1.2.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
Faker==19.13.0
fastapi==0.68.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
lxml==5.3.0
lxml_html_clean==0.4.1
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
psycopg2-binary==2.9.9
pydantic==1.10.20
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
requests==2.26.0
requests-html==0.10.0
selenium==4.9.0
selenium-stealth==1.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.14.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==1.26.20
uvicorn==0.15.0
w3lib==2.2.1
webdriver-manager==4.0.1
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
